refactor(guest-mock-claude): split mock cli modules
